Validate

+++ app/Http/Controllers/Api/V1/ContragentsController.php
@@ -4,11 +4,16 @@ namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 use \App\Contragent;
+use Illuminate\Support\Facades\Auth;
 use \App\Store;
+use Illuminate\Support\Facades\Validator;
 
 class ContragentsController extends Controller
 {
+
+
     /**
      * Display a listing of the resource.
      *
@@ -27,6 +32,8 @@ class ContragentsController extends Controller
      */
     public function store(Request $request)
     {
+
+
         $contragent = Contragent::create($request->all());
         $contragent->types()->sync($request->all()['typeIds']);
         $storesIds = [];
@@ -71,6 +78,75 @@ class ContragentsController extends Controller
         return Contragent::findOrFail($id);
     }
 
+    public function my()
+    {
+        return Contragent::findOrFail(Auth::user()->contragents[0]->id);
+    }
+
+    private function checkUser($id){
+
+        if(Auth::user()->role_id == 1) return true;
+        if(Auth::user()->contragents[0]->id !== $id) return false;
+        return true;
+
+    }
+
+
+    public function update(Request $request, $id)
+    {
+
+        $validator = Validator::make($request->all(), [
+            'inn' => [
+                'required',
+                Rule::unique('contragents')->ignore($id),
+            ],
+            'title' => [
+                'required'
+            ],
+            'legal_address' => [
+                'required'
+            ],
+        ]);
+
+        if(!$validator->fails() && $this->checkUser($id)){
+
+            $contragent = Contragent::findOrFail($id);
+            $contragent->types()->sync($request->all()['typeIds']);
+            $storesIds = [];
+            $contragent->update($request->all());
+            if(count($request->all()['stores'])){
+                foreach($request->all()['stores'] as $store){
+                    if($store['coords'] && $store['address']){
+                        if($store['id']){
+                            Store::find($store['id'])->update([
+                                'contragent_id' => $contragent->id,
+                                'coords' => $store['coords'], 
+                                'address' => $store['address'], 
+                                'federal_district_id' => $store['federal_district']['id'],
+                                'region_id' => $store['region']['id']
+                            ]);
+                            $storesIds[] = $store['id'];
+                        } else {
+                            $storesIds[] = Store::create([
+                                'contragent_id' => $contragent->id,
+                                'coords' => $store['coords'],
+                                'address' => $store['address'],
+                                'federal_district_id' => $store['federal_district']['id'],
+                                'region_id' => $store['region']['id']
+                            ])->id;
+                        }
+                    }
+                }
+                Store::whereNotIn('id', $storesIds)->where("contragent_id", $contragent->id)->delete();
+            }
+        }
+
+        $contragent = Contragent::findOrFail($id);
+        $contragent->errors = $validator->errors();
+        return $contragent;
+    }
+
+
     /**
      * Update the specified resource in storage.
      *
@@ -78,38 +154,58 @@ class ContragentsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function updateCompany(Request $request)
     {
-        $contragent = Contragent::findOrFail($id);
-        $contragent->types()->sync($request->all()['typeIds']);
-        $storesIds = [];
-        $contragent->update($request->all());
-        if(count($request->all()['stores'])){
-            foreach($request->all()['stores'] as $store){
-                if($store['coords'] && $store['address']){
-                    if($store['id']){
-                        Store::find($store['id'])->update([
-                            'contragent_id' => $contragent->id,
-                            'coords' => $store['coords'], 
-                            'address' => $store['address'], 
-                            'federal_district_id' => $store['federal_district']['id'],
-                            'region_id' => $store['region']['id']
-                        ]);
-                        $storesIds[] = $store['id'];
-                    } else {
-                        $storesIds[] = Store::create([
-                            'contragent_id' => $contragent->id,
-                            'coords' => $store['coords'],
-                            'address' => $store['address'],
-                            'federal_district_id' => $store['federal_district']['id'],
-                            'region_id' => $store['region']['id']
-                        ])->id;
+        $id = Auth::user()->contragents[0]->id;
+
+        $validator = Validator::make($request->all(), [
+            'inn' => [
+                'required',
+                Rule::unique('contragents')->ignore($id),
+            ],
+            'title' => [
+                'required'
+            ],
+            'legal_address' => [
+                'required'
+            ],
+        ]);
+
+        if(!$validator->fails()){
+
+            $contragent = Contragent::findOrFail($id);
+            $contragent->types()->sync($request->all()['typeIds']);
+            $storesIds = [];
+            $contragent->update($request->all());
+            if(count($request->all()['stores'])){
+                foreach($request->all()['stores'] as $store){
+                    if($store['coords'] && $store['address']){
+                        if($store['id']){
+                            Store::find($store['id'])->update([
+                                'contragent_id' => $contragent->id,
+                                'coords' => $store['coords'], 
+                                'address' => $store['address'], 
+                                'federal_district_id' => $store['federal_district']['id'],
+                                'region_id' => $store['region']['id']
+                            ]);
+                            $storesIds[] = $store['id'];
+                        } else {
+                            $storesIds[] = Store::create([
+                                'contragent_id' => $contragent->id,
+                                'coords' => $store['coords'],
+                                'address' => $store['address'],
+                                'federal_district_id' => $store['federal_district']['id'],
+                                'region_id' => $store['region']['id']
+                            ])->id;
+                        }
                     }
                 }
+                Store::whereNotIn('id', $storesIds)->where("contragent_id", $contragent->id)->delete();
             }
-            Store::whereNotIn('id', $storesIds)->where("contragent_id", $contragent->id)->delete();
         }
+
         $contragent = Contragent::findOrFail($id);
+        $contragent->errors = $validator->errors();
         return $contragent;
     }
 
@@ -121,8 +217,8 @@ class ContragentsController extends Controller
      */
     public function destroy($id)
     {
-        $contragent = Contragent::findOrFail($id);
-        $contragent->delete();
+        // $contragent = Contragent::findOrFail($id);
+        // $contragent->delete();
         return '';
     }
 }
